ft(refactor-response-object):refactor the response object to use JSEND format - refactor response object to use JSEND format in server/controller/requestDbController.js - refactor response object to use JSEND format in server/controller/userDbController.js - refactor response object to JSEND format server/middleware/authenticate.js - remove utility functions sendSuccess and sendMessage from server/middleware/findRequestById.js - create utils folder to hold utility functions in server/utils/utils.js - rename server/middleware/utils.js to server/middleware/findRequestById.js - modify test to reflect changes in response object
