+++ .travis.yml
@@ -1,3 +1,7 @@
+env:
+  global:
+    - CC_TEST_REPORTER_ID=6e4f39ecc1de8f4993f93ffc47c768bad33aa37fc307a8d2b4e79af6a9386041
+
 language: node_js
 node_js:
   - "6"
@@ -9,4 +13,13 @@ services:
 
 before_script:
   - psql -c 'create database travis;' -U postgres
+  - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
+  - chmod +x ./cc-test-reporter
+  - ./cc-test-reporter before-build
+
+script:
+  - npm run test
+
+after_script:
+  - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
